Misc tweaks and fixes
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@23458 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/render3d/opengl/gl_render3d.h
@@ -260,11 +260,11 @@ private:
   //Calculate if the portal clipportal_stack[index] is
   //mirrored by preceding portals. Each reflection reverses
   //vertex order of portal polygon.
-  bool IsPortalMirrored(int index)
+  bool IsPortalMirrored(size_t index)
   {
     bool mirror = false;
-    for (int i=0;i<index;i++) 
-      if (clipportal_stack[i]->flags.Check(CS_OPENPORTAL_MIRROR)) 
+    for (size_t i = 0; i < index; i++) 
+      if (clipportal_stack[i]->flags.Check (CS_OPENPORTAL_MIRROR)) 
         mirror = !mirror;
     return mirror;
   }
